+++ app/src/main/java/app/revanced/integrations/youtube/patches/components/ShortsFilter.java
@@ -21,13 +21,17 @@ public final class ShortsFilter extends Filter {
     private final ByteArrayFilterGroup shortsCompactFeedVideoBuffer;
 
     private final StringFilterGroup channelBar;
+    private final StringFilterGroup fullVideoLinkLabel;
+    private final StringFilterGroup videoTitle;
+    private final StringFilterGroup reelSoundMetadata;
     private final StringFilterGroup subscribeButton;
     private final StringFilterGroup subscribeButtonPaused;
     private final StringFilterGroup soundButton;
     private final StringFilterGroup infoPanel;
+    private final StringFilterGroup joinButton;
     private final StringFilterGroup shelfHeader;
 
-    private final StringFilterGroup videoActionButton;
+    private final StringFilterGroup actionBar;
     private final ByteArrayFilterGroupList videoActionButtonGroupList = new ByteArrayFilterGroupList();
 
     public ShortsFilter() {
@@ -64,7 +68,7 @@ public final class ShortsFilter extends Filter {
                 "/frame0.jpg");
 
         // Shorts player components.
-        var joinButton = new StringFilterGroup(
+        joinButton = new StringFilterGroup(
                 Settings.HIDE_SHORTS_JOIN_BUTTON,
                 "sponsor_button"
         );
@@ -84,6 +88,21 @@ public final class ShortsFilter extends Filter {
                 REEL_CHANNEL_BAR_PATH
         );
 
+        fullVideoLinkLabel = new StringFilterGroup(
+                Settings.HIDE_SHORTS_FULL_VIDEO_LINK_LABEL,
+                "reel_multi_format_link"
+        );
+
+        videoTitle = new StringFilterGroup(
+                Settings.HIDE_SHORTS_VIDEO_TITLE,
+                "shorts_video_title_item"
+        );
+
+        reelSoundMetadata = new StringFilterGroup(
+                Settings.HIDE_SHORTS_SOUND_METADATA_LABEL,
+                "reel_sound_metadata"
+        );
+
         soundButton = new StringFilterGroup(
                 Settings.HIDE_SHORTS_SOUND_BUTTON,
                 "reel_pivot_button"
@@ -94,15 +113,26 @@ public final class ShortsFilter extends Filter {
                 "shorts_info_panel_overview"
         );
 
-        videoActionButton = new StringFilterGroup(
+        actionBar = new StringFilterGroup(
                 null,
-                "ContainerType|shorts_video_action_button"
+                "shorts_action_bar"
         );
 
         addPathCallbacks(
                 shortsCompactFeedVideoPath,
                 joinButton, subscribeButton, subscribeButtonPaused,
-                channelBar, soundButton, infoPanel, videoActionButton
+                channelBar, fullVideoLinkLabel, videoTitle, reelSoundMetadata,
+                soundButton, infoPanel, actionBar
+        );
+
+        var shortsLikeButton = new ByteArrayFilterGroup(
+                Settings.HIDE_SHORTS_LIKE_BUTTON,
+                "shorts_like_button"
+        );
+
+        var shortsDislikeButton = new ByteArrayFilterGroup(
+                Settings.HIDE_SHORTS_DISLIKE_BUTTON,
+                "shorts_dislike_button"
         );
 
         var shortsCommentButton = new ByteArrayFilterGroup(
@@ -120,7 +150,13 @@ public final class ShortsFilter extends Filter {
                 "reel_remix_button"
         );
 
-        videoActionButtonGroupList.addAll(shortsCommentButton, shortsShareButton, shortsRemixButton);
+        videoActionButtonGroupList.addAll(
+                shortsLikeButton,
+                shortsDislikeButton,
+                shortsCommentButton,
+                shortsShareButton,
+                shortsRemixButton
+        );
     }
 
     @Override
@@ -131,6 +167,9 @@ public final class ShortsFilter extends Filter {
             if (matchedGroup == soundButton ||
                     matchedGroup == infoPanel ||
                     matchedGroup == channelBar ||
+                    matchedGroup == fullVideoLinkLabel ||
+                    matchedGroup == videoTitle ||
+                    matchedGroup == reelSoundMetadata ||
                     matchedGroup == subscribeButtonPaused
             ) return super.isFiltered(identifier, path, protobufBufferArray, matchedGroup, contentType, contentIndex);
 
@@ -141,8 +180,8 @@ public final class ShortsFilter extends Filter {
                 return false;
             }
 
-            // Video action buttons (comment, share, remix) have the same path.
-            if (matchedGroup == videoActionButton) {
+            // Video action buttons (like, dislike, comment, share, remix) have the same path.
+            if (matchedGroup == actionBar) {
                 if (videoActionButtonGroupList.check(protobufBufferArray).isFiltered()) return super.isFiltered(
                         identifier, path, protobufBufferArray, matchedGroup, contentType, contentIndex
                 );
@@ -151,7 +190,9 @@ public final class ShortsFilter extends Filter {
 
             // Filter other path groups from pathFilterGroupList, only when reelChannelBar is visible
             // to avoid false positives.
-            if (matchedGroup == subscribeButton) {
+            if (matchedGroup == subscribeButton ||
+                    matchedGroup == joinButton
+            ) {
                 if (path.startsWith(REEL_CHANNEL_BAR_PATH)) return super.isFiltered(
                         identifier, path, protobufBufferArray, matchedGroup, contentType, contentIndex
                 );